Add functions to populate the display
The user can digits multiple operators and numbers, whose operation is executed
on the go.
Since the back button is possible, there are x1, x2, operator which are arrays,
rather than simple variables.
The current values to be used are in the first position (so that it is easier to
reference them), while the oldest values are stored in the next indices.
Note that x1, except for the very first number that is digited, is the result of
the previous operations.

Variable digitStatus helps distinguish whether we are digiting a number or an
operator, and its index.

Note: some buttons (e.g., '.', '%', '^') are not functional yet.